Initial Sprint
Works, but a bit rough on the edges.
…
And the middle, like when you run it
without any parameter nor input at all.

It'll expect you to type your CSV and
hit CTRL+D (detach).

Best make sure you give it parameters then.

Still, we import no libs but standard,
and that's good for portability.